ftp_server/requirements.txt
pyftpdlib==1.5.6
kafka-python==2.0.2
boto3==1.21.*
